Moved PnP optimization out of experimental folder
