+++ rct_optimizations/src/rct_optimizations/pnp.cpp
@@ -1,4 +1,4 @@
-#include "rct_optimizations/experimental/pnp.h"
+#include "rct_optimizations/pnp.h"
 #include "rct_optimizations/ceres_math_utilities.h"
 #include "rct_optimizations/eigen_conversions.h"
